+++ feedback_api/src/main/java/feedback_flow/feedback_api/application/repositories/CustomerFeedbackRepository.java
@@ -1,0 +1,4 @@
+package feedback_flow.feedback_api.application.repositories;
+
+public class CustomerFeedbackRepository {
+}
