+++ feedback_api/src/main/java/feedback_flow/feedback_api/application/services/CustomerFeedbackService.java
@@ -1,0 +1,4 @@
+package feedback_flow.feedback_api.application.services;
+
+public class CustomerFeedbackService {
+}
