--- feedback_api/src/main/java/feedback_flow/feedback_api/domain/repository/CustomerFeedback.java
@@ -1,4 +1,0 @@
-package feedback_flow.feedback_api.domain.repository;
-
-public class CustomerFeedback {
-}
