--- feedback_api/src/main/java/feedback_flow/feedback_api/domain/service/CustomerFeedback.java
@@ -1,4 +1,0 @@
-package feedback_flow.feedback_api.domain.service;
-
-public class CustomerFeedback {
-}
